MultiGridBarrier

Documentation for MultiGridBarrier.

MultiGridBarrier.AMG

MultiGridBarrier.Barrier

MultiGridBarrier.amg

MultiGridBarrier.amgb

MultiGridBarrier.barrier

MultiGridBarrier.damped_newton

MultiGridBarrier.fem1d

MultiGridBarrier.fem2d

MultiGridBarrier.fem_interp1d

MultiGridBarrier.fem_interp1d

MultiGridBarrier.fem_plot_2d

MultiGridBarrier.fem_solve_1d

MultiGridBarrier.fem_solve_2d

MultiGridBarrier.interp1d

MultiGridBarrier.interp2d

MultiGridBarrier.plot1d

MultiGridBarrier.plot2d

MultiGridBarrier.spectral_solve_1d

MultiGridBarrier.spectral_solve_2d

MultiGridBarrier.AMG — Type

@kwdef struct AMG{T,M}
    ...
end

Objects of this type should probably be assembled by the constructor amg().

A multigrid with L level. Denote by l between 1 and L, a grid level. Fields are:

x::Array{Array{T,2},1} an array of L matrices. x[l] stores the vertices of the grid at multigrid level l.

w::Array{Array{T,1},1} an array of L quadrature weights. w[l] corresponds to x[l].

R_fine::Array{M,1} an array of L matrices. The columns of R_fine[l] are basis functions for the function space on grid level l, interpolated to the fine grid.

R_coarse::Array{M,1} an array of L matrices. The columns of R_coarse[l] are basis functions for the function space on grid level l. Unlike R_fine[l], these basis functions are on grid level l, not interpolated to the fine grid.

D::Array{M,2} an array of differential operators. For example, if the barrier parameters are to be u,ux,s, with ux the derivative of u, then D[l,:] = [I,Dx,I], where Dx is a numerical differentiation operator on grid level l.

refine_u::Array{M,1} an array of L grid refinement matrices. If x[l] has n[l] vertices, then refine_u[l] is n[l+1] by n[l].

coarsen_u::Array{M,1} an array of L grid coarsening matrices. coarsen_u[l] is n[l] by n[l+1].

refine_z::Array{M,1} an array of L grid refining matrices for the "state vector" z. For example, if z contains the state functions u and s, then there are k=2 state functions, and refine_z[l] is k*n[l+1] by k*n[l].

coarsen_z::Array{M,1} an array of L grid coarsening matrices for the "state vector" z. coarsen_z[l] is k*n[l] by k*n[l+1].

These various matrices must satisfy a wide variety of algebraic relations. For this reason, it is recommended to use the constructor amg().

MultiGridBarrier.Barrier — Type

Barrier

A type for holding barrier functions. Fields are:

f0::Function
f1::Function
f2::Function

f0 is the barrier function itself, while f1 is its gradient and f2 is the Hessian.

MultiGridBarrier.amg — Method

function amg(;
    x::Array{Array{T,2},1},
    w::Array{T,1},
    state_variables::Array{Symbol,2},
    D::Array{Symbol,2},
    subspaces::Dict{Symbol,Array{M,1}},
    operators::Dict{Symbol,M},
    refine::Array{M,1},
    coarsen::Array{M,1}) where {T,M}

Construct an AMG object for use with the amgb solver. In many cases, this constructor is not called directly by the user. For 1d and 2d finite elements, use the MultiGridBarrier module. For 1d and 2d spectral elements, use the SpectralBarrierMethod module. You use amg() directly if you are implementing your own function spaces.

The AMG object shall represent all L grid levels of the multigrid hierarchy. Parameters are:

x: an array of L matrices. x[l] has the vertices of grid level l, one vertex per row.

w: an array of L vectors. w[l] has the quadrature weights for grid level l.

state_variables: a matrix of symbols. The first column indicates the names of the state vectors or functions, and the second column indicates the names of the corresponding subspaces. A typical example is: state_variables = [:u :dirichlet; :s :full]. This would define the solution as being functions named u(x) and s(x). The u function would lie in the space :dirichlet, presumably consisting of functions with homogeneous Dirichlet conditions. The s function would lie in the space :full, presumably being the full function space, without boundary conditions.

D: a matrix of symbols. The first column indicates the names of various state variables, and the second column indicates the corresponding differentiation operator(s). For example: D = [:u :id ; :u :dx ; :s :id]. This would indicate that the barrier should be called as F(x,y) with y = [u,ux,s], where ux denotes the derivative of u with respect to the space variable x.

subspaces: a Dict mapping each subspace symbol to an array of L matrices, e.g. for each l, subspaces[:dirichlet][l] is a matrix whose columns span the homogeneous Dirichlet subspace of grid level l.

operators: a Dict mapping each differential operator symbol to a matrix, e.g. operators[:id] is an identity matrix, while operators[:dx] is a numerical differentiation matrix, on the fine grid level L.

refine: an array of length L of matrices. For each l, refine[l] interpolates from grid level l to grid level l+1. refine[L] should be the identity, and coarsen[l]*refine[l] should be the identity.

coarsen: an array of length L of matrices. For each l, coarsen[l] interpolates or projects from grid level l+1 to grid level l. coarsen[L] should be the identity.

MultiGridBarrier.amgb — Method

function amgb(B::Barrier,
    M::AMG{T,Mat},
    z::Array{T,1},
    c::Array{T,2};
    tol=(eps(T)),
    t=T(0.1),
    maxit=10000,
    alpha=T(0.5),
    beta=T(0.1),
    kappa=T(10.0),
    verbose=true) where {T,Mat}

The "Algebraic MultiGrid Barrier" method.

B a Barrier object.

M an AMG object.

z a starting point for the minimization, which should be admissible, i.e. B.f0(z)<∞.

c an objective functional to minimize. Concretely, we minimize the integral of c.*(D*z), as computed by the finest quadrature in M, subject to B.f0(z)<∞. Here, D is the differential operator provided in M.

Optional parameters:

t: the initial value of t

tol: we stop when 1/t<tol.

maxit: the maximum number of t steps.

alpha, beta: parameters of the backtracking line search.

kappa: the initial size of the t-step. Stepsize adaptation is used in the AMGB algorithm, where the t-step size may be made smaller or large, but it will never exceed the initial size provided here.

verbose: set to true to see a progress bar.

Return value is a named tuple SOL with the following fields:

SOL.converged is true if convergence was obtained, else it is false.

SOL.z the computed solution.

Further SOL fields contain various statistics about the solve process.

The following "example usage" is an extremely convoluted way of minimizing x in the interval [-1,1]:

using AlgebraicMultiGridBarrier
M = amg(x = [[-1.0 ; 1.0 ;;]],
        w = [1.0,1.0],
        state_variables = [:u :space],
        D = [:u :id],
        subspaces = Dict(:space => [[1.0 ; -1.0 ;;]]),
        operators = Dict(:id => [1.0 0.0;0.0 1.0]),
        refine = [[1.0 0.0 ; 0.0 1.0]],
        coarsen = [[1.0 0.0 ; 0.0 1.0]])
B = barrier((x,y)->-log(1-x[1]*y[1]))
amgb(B,M,[0.0,0.0],[1.0 ; 0.0 ;;])

MultiGridBarrier.barrier — Method

function barrier(f;
    f1=(x,y)->ForwardDiff.gradient(z->f(x,z),y),
    f2=(x,y)->ForwardDiff.hessian(z->f(x,z),y))::Barrier

Constructor for barriers.

f is the actual barrier function. It should take parameters (x,y).

f1 is the gradient of f with respect to y.

f2 is the Hessian of f with respect to y.

By default, f1 and f2 are automatically generated by the module ForwardDiff.

MultiGridBarrier.damped_newton — Method

function damped_newton(F0::Function,
                   F1::Function,
                   F2::Function,
                   x::Array{T,1};
                   maxit=10000,
                   alpha=T(0.5),
                   beta=T(0.1)) where {T}

Damped Newton iteration for minimizing a function.

F0 the objective function

F1 and F2 are the gradient and Hessian of F0, respectively.

x the starting point of the minimization procedure.

The optional parameters are:

maxit, the iteration aborts with a failure message if convergence is not achieved within maxit iterations.

alpha and beta are the parameters of the backtracking line search.

tol is used as a stopping criterion. We stop when the decrement in the objective is sufficiently small.

MultiGridBarrier.fem1d — Method

function fem1d(::Type{T}, L::Int;
                state_variables = [:u :dirichlet
                                   :s :full],
                D = [:u :id
                     :u :dx
                     :s :id]) where {T}

Construct an AlgebraicMultiGridBarrier.AMG object for a 1d piecewise linear finite element grid. The interval is [-1,1]. Parameters are:

L: divide the interval into 2^L subintervals (L for Levels).

state_variables: the "state vector" consists of functions, by default this is u(x) and s(x), on the finite element grid.

D: the set of differential operator. The barrier function F will eventually be called with the parameters F(x,Dz), where z is the state vector. By default, this results in F(x,u,ux,s), where ux is the derivative of u.

MultiGridBarrier.fem2d — Method

function fem2d(::Type{T}, L::Int, K::Matrix{T};
                state_variables = [:u :dirichlet
                                   :s :full],
                D = [:u :id
                     :u :dx
                     :u :dy
                     :s :id]) where {T}

Construct an AlgebraicMultiGridBarrier.AMG object for a 2d finite element grid on the domain K with piecewise quadratic elements. Parameters are:

K: a triangular mesh. If there are n triangles, then K should be a 3n by 2 matrix of vertices. The first column of K represents x coordinates, the second column represents y coordinates.

L: divide the interval into 2^L subintervals (L for Levels).

state_variables: the "state vector" consists of functions, by default this is u(x) and s(x), on the finite element grid.

D: the set of differential operator. The barrier function F will eventually be called with the parameters F(x,y,Dz), where z is the state vector. By default, this results in F(x,y,u,ux,uy,s), where (ux,uy) is the gradient of u.

MultiGridBarrier.fem_interp1d — Method

function fem_interp1d(x::Vector{T},
                  y::Vector{T},
                  t::T) where{T}

Interpolate a 1d piecewise linear function at the given t value. If u(xi) is the piecewise linear function such that u(x[k])=y[k] then this function returns u(t).

MultiGridBarrier.fem_interp1d — Method

function fem_interp1d(x::Vector{T}, y::Vector{T}, t::Vector{T}) where{T}

Returns [fem_interp1d(x,y,t[k]) for k=1:length(t)].

MultiGridBarrier.fem_plot_2d — Method

function fem_plot_2d(M::AMG{T, Mat}, z::Array{T}) where {T,Mat}

Plot a piecewise quadratic solution z on the given mesh. Note that the solution is drawn as (linear) triangles, even though the underlying solution is piecewise quadratic. To obtain a more accurate depiction, especially when the mesh is coarse, it would be preferable to apply a few levels of additional subdivision, so as to capture the curve of the quadratic basis functions.

MultiGridBarrier.fem_solve_1d — Method

function fem_solve_1d(::Type{T}; g = x->x,
    f = x->T(0.5), maxit=10000, L=2, p=T(1.0),
    verbose=true, show=true, tol=sqrt(eps(T)),
    F = (x,u,ux,s) -> -log(s^(2/p)-ux^2)-2*log(s),
    slack = x->T(2)) where {T}

Solve a 1d variational problem on the interval [-1,1] with piecewise linear elements. Parameters are:

g the boundary conditions.

f the forcing function.

maxit a maximum number of iterations used in the solver.

L the number of Levels of grid subdivisions, so that the grid consists of 2^L intervals.

p the parameter of the p-Laplace problem, if that's what we're solving.

verbose: set to true to get a progress bar.

tol: a stopping criterion, the barrier method stops when t>1/tol.

F: the barrier. The default barrier solves a p-Laplacian.

slack: an initializer for the slack function s(x).

This function returns SOL,B, where SOL is from amgb, and B is the Barrier object obtained from F.

MultiGridBarrier.fem_solve_2d — Method

function fem_solve_2d(::Type{T}; 
    K = T[-1 -1;1 -1;-1 1;1 -1;1 1;-1 1],
    g = (x,y)->x^2+y^2, 
    f = (x,y)->T(0.5), maxit=10000, L=2, p=T(1.0),
    verbose=true, show=true, tol=sqrt(eps(T)),
    F = (x,y,u,ux,uy,s) -> -log(s^(2/p)-ux^2-uy^2)-2*log(s),
    slack = (x,y)->T(100)) where {T}

Solve a 2d variational problem on the domain K, which defaults to the square [-1,1]x[-1,1], with piecewise quadratic elements. Parameters are:

K a triangulation of the domain. For n triangles, K should be a 3n by 2 matrix of vertices.

g the boundary conditions.

f the forcing function.

maxit a maximum number of iterations used in the solver.

L the number of Levels of grid subdivisions, so that the grid consists of N = n*4^L quadratic triangular elements. Each elements is quadratic, plus a bump function, so each element consists of 7 vertices, i.e. there are 7*N vertices in total.

p the parameter of the p-Laplace problem, if that's what we're solving.

verbose: set to true to get a progress bar.

tol: a stopping criterion, the barrier method stops when t>1/tol.

F: the barrier. The default barrier solves a p-Laplacian.

slack: an initializer for the slack function s(x).

This function returns SOL,B, where SOL is from amgb, and B is the Barrier object obtained from F.

MultiGridBarrier.interp1d — Method

function interp1d(MM::AMG{T,Mat}, y::Array{T,1},x) where {T,Mat}

A function to interpolate a solution y at some point(s) x.

MM the mesh of the solution.

y the solution.

x point(s) at which the solution should be evaluated.

MultiGridBarrier.interp2d — Method

function interp2d(MM::AMG{T,Mat},z::Array{T,1},x::Array{T,2}) where {T,Mat}

Interpolate a solution z at point(s) x, given the mesh MM. See also interp1d.

MultiGridBarrier.plot1d — Method

function plot1d(M::AMG{T,Mat},x,y,rest...) where {T,Mat}

Plot a solution using pyplot.

M: a mesh.

x: x values where the solution should be evaluated and plotted.

y: the solution, to be interpolated at the given x values via interp1d.

rest... parameters are passed directly to pyplot.plot.

MultiGridBarrier.plot2d — Method

function plot2d(M::Mesh{T},x,y,z::Array{T,1};rest...) where {T}

Plot a 2d solution.

M a 2d mesh.

x, y should be ranges like -1:0.01:1.

z the solution to plot.

MultiGridBarrier.spectral_solve_1d — Method

function spectral_solve_1d(::Type{T}; g = x->x,
    f = x->T(0.5), maxit=10000, n=4, p=T(1.0),
    verbose=true, show=true, tol=sqrt(eps(T)),
    F = (x,u,ux,s) -> -log(s^(2/p)-ux^2)-2*log(s),
    slack = x->T(2)) where {T}

Solves a p-Laplace problem in d=1 dimension with the given value of p and plot the result.

MultiGridBarrier.spectral_solve_2d — Method

function spectral_solve_2d(::Type{T}; g = (x,y)->x^2+y^2, 
    f = (x,y)->T(0.5), maxit=10000, n=4, p=T(1.0),
    verbose=true, show=true, tol=sqrt(eps(T)),
    F = (x,y,u,ux,uy,s) -> -log(s^(2/p)-ux^2-uy^2)-2*log(s),
    slack = (x,y)->T(10)) where {T}

Solves a p-Laplace problem in d=2 dimensions with the given value of p and plot the result.